% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bcemtsubtype.R
\name{predict_bcemtsubtype}
\alias{predict_bcemtsubtype}
\title{Breast Cancer EMT Subtype classifier}
\usage{
predict_bcemtsubtype(expr.dat, scale = T)
}
\arguments{
\item{expr.dat}{An object of class matrix with samples as column names, gene symbols as row names, and entries the corresponding expression levels.}

\item{scale}{logical, whether to scale-normalization for data.}
}
\value{
A dataframe  with the prediction results.
}
\description{
A function to predict the subtypes of breast cancer
based on EMT with respect to a ML model.
}
